Fix checkstyle dependency and run it sooner. (#66)
`getTaskName` returns `checkstyle` for the `main` source set, but the checkstyle task is `checkstyleMain`. So the dependency from `checkstyle` to `checkstyleMain` isn't being made (`checkstyleMain` still runs in the build because the checkstyle plugin adds the dependency on it from `check`).

Also made it so checkstyle runs just after compiling to get feedback earlier.

These changes make it so instead of `:core:checkstyleMain` being one of the last things run after 30 min on Travis to one of the first as it probably should be.