Add `lint` and `generateSources` task
Motivation:

`checkstyle` is a tool-specific task name. There has to be a generic
task name for running all linting tools, so that a user can easily
run all the necessary linting tasks for the entire codebase.

There's no dedicated task for populating the generated source files. A
user had to rely on the tasks such as `checkstyle` so far.

Modifications:

- Added the `lint` task which is supposed to run all linting tasks,
  including `checkstyle`.
- Added the `generateSources` task which is supposed to run all
  source-generating tasks, including `compile*Thrift` and `generate*Proto`.
- Deprecated `noCheckstyle` property in favor of `noLint`.

Result:

- The new `lint` and `generateSources` tasks allow a user to add other
  linters/generators than Checkstyle/Protobuf/Thrift into the build
  process easily.
- We now can tell a user to run `generateSources` task before importing
  a project into an IDE.