Wait for process output when executing commands (#91)
#### Motivation
When using `Process.waitFor()` and executing a command with a lot of output (for example `git tag -l "foo*"`), the stream buffers can fill up, causing the program to hang, waiting for user to read the streams. This seems to depend on OS implementation, which made it difficult to replicate.

Using `Process.waitForProcessOutput()` fixes this issue since it consumes stdout and stderr in separate threads before waiting for the process to exit.

#### Modifications
- Use `Process.waitForProcessOutput()` instead of `Process.waitFor()`
- Trim final newline in error message

#### Result
Commands with a lot of output do not hang anymore.