Fix a bug where `dependencyUpdates` task does not work correctly with `dependencies.toml`
Motivation:

The `dependencyUpdates` task is configured in
`dependencyManagementProject` does not report new versions in
Maven Central.
https://github.com/line/gradle-scripts/blob/master/lib/common-dependencies.gradle#L56
I don't know the exact reason of that but `dependencyUpdates` task is
worked correctly when it is configured in the root project guided in
`gradle-versions-plugin` documentation.
https://github.com/ben-manes/gradle-versions-plugin#multi-project-build

Modifications:

- Upgrade gradle-versions-plugin to 0.42.0 from 0.39.0
- Move `dependencyUpdates` configuration to `rootProject`.
- Add a custom reporter so as to print Maven Search urls with all new
  versions together.
  - It would be helpful to check legacy versions.
    `managedDependencyOverrides` is used to provide the URL but it no
    longer works with version catalog.

Result:

You can now correctly see new versions with `dependencyUpdates` task
when using `dependencies.toml`