Omit dependencyManagement module from direct dependency (#165)
Motivation:

As `dependencyManagement` is directly added as a project dependency to all projects, it is listed in the POM file and Gradle module metadata. Consequently, the `dependencyManagement` module is manually removed from the POM dependencies. However, there is no extension point to customize the Gradle module metadata file.

It would be a easier way to remove `dependencyManagement` from the project dependency fundamentally rather than removing later when publishing.

Modifications:

- Create `internal` configuration and add `dependencyManagement` dependency to it.
- `internal` configuration is set as the parent configuration of `*Classpath`, shaded, and Proguard configurations.
- Removed the legacy configuration format `dependencies.yaml` to remove maintenance costs

Result:

- `gradle-scripts` correctly produces Gradle module metadata if no `trim` or `shade` flags are set
- ⚠️ `dependencies.yaml` is no longer supported. Please use `dependencies.toml` instead.