Remove invalid access to List.value in bom.gradle (#173)
`subs` should be an ArrayList got from `bomGroups`, but this script somehow accesses to its property `subs.value`.

We confirmed with maintainers that this is a bug. This `subs.value` property happens to be a list before Java 17, but in Java 17 it becomes an invalid reference.

```
* What went wrong:
A problem occurred configuring project ':foo-project:bom'.
> Exception evaluating property 'value' for java.util.ArrayList, Reason: groovy.lang.MissingPropertyException: No such property: value for class: java.lang.String
```

So this PR gets rid of these references.